pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "teledex"
version = "0.1.0"
description = "Bridge a local Codex CLI session to Telegram on macOS."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=22.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
faster-whisper = [
  "faster-whisper>=1.1.1",
]

[project.scripts]
teledex = "teledex.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
